+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 0.1.0 - 2017-08-07
 
+- `setKeyValue` & `setKeyValueSync` implemented
 - `deleteKey` & `deleteKeySync` implemented
-- Start of the project
+- Start of the library